Refactor to unified metabo cycle
